fix : searchBar 관련 태그 및 컴포넌트 위치 수정
