+++ src/main/front/src/components/SearchResult.js
@@ -1,6 +1,7 @@
 import React from "react";
 
-import styles from "../styles/Header.module.css";
+//CSS
+import styles from "../styles/SearchBar.module.css";
 import { FaHashtag } from "react-icons/fa";
 
 function SearchResult(props) {
