+++ src/main/front/src/styles/reset.css
@@ -12,3 +12,11 @@ input {
   box-shadow: none;
   background: none;
 }
+
+a {
+  text-decoration: none;
+}
+
+a:active {
+  color: black;
+}
